AI embeddings: extract the transformers runner body to runners/torch_embed.py
Mirrors the diffusers_image/torch_image.py split (D381): a hardware-variant
folder's worker.py must be a five-line shell so a second copy of the model
logic never has the chance to drift, and the ROCm/CUDA rows this branch adds
next both need that shell to import. transformers_embed/worker.py keeps its
own pyproject.toml but now imports runners/torch_embed.py for everything else,
byte-for-byte the same download/_placement/load/_pooled/generate/main it had
before extraction — _pooled and its pooler_output docstring (just fixed in
9ffcf768) survive unchanged.

tests/test_ai_transformers_embed_worker.py now targets torch_embed.py
directly rather than the transformers_embed/ shell, the same choice
test_ai_diffusers_worker.py already makes for torch_image.py.
